Implemented POST request API  for registering a new user and storing details to ShowtimeDB. (#6)
* Creating generic classes with folder structure and added README file.

* Handling comments and some minor changes.

* Implemented POST API for registering a new user and storing details to
DB.

---------

+++ src/main/java/com/application/showtime/UserOnboarding/repository/UserRegistrationRepo.java
@@ -5,5 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserRegistrationRepo extends CrudRepository<UserRegistrationEntity,Integer> {
+public interface UserRegistrationRepo extends CrudRepository<UserRegistrationEntity,Long> {
+
 }
